[test] split parser tests into individual files
