+++ test/unit/test_json_parser.py
@@ -122,89 +122,3 @@ class TestJSONParser(object):
         assert_equal(set(parsed_data[0].keys()), {'name', 'age', 'city', 'state'})
         assert_equal(parsed_data[0]['name'], 'john')
         assert_equal(type(parsed_data[0]['age']), int)
-
-class TestCloudWatchParser(object):
-    @classmethod
-    def setup_class(cls):
-        """setup_class() before any methods in this class"""
-        pass
-
-    @classmethod
-    def teardown_class(cls):
-        """teardown_class() after any methods in this class"""
-        pass
-
-    def setup(self):
-        """Setup before each method"""
-        # load config
-        self.config = load_config('test/unit/conf')
-        # load JSON parser class
-        self.parser_class = get_parser('gzip-json')
-
-    def teardown(self):
-        """Teardown after each method"""
-        pass
-
-    def parser_helper(self, **kwargs):
-        data = kwargs['data']
-        schema = kwargs['schema']
-        options = kwargs['options']
-
-        json_parser = self.parser_class(data, schema, options)
-        parsed_result = json_parser.parse()
-        return parsed_result
-
-    def test_cloudwatch(self):
-        """Parse CloudWatch JSON"""
-        schema = self.config['logs']['test_cloudwatch']['schema']
-        options = { "hints": self.config['logs']['test_cloudwatch']['hints']}
-        with open('test/unit/fixtures/cloudwatch.json','r') as fixture_file:
-            data = fixture_file.readlines()
-        data_record = zlib.compress(data[0].strip())
-        parsed_result = self.parser_helper(data=data_record, schema=schema, options=options)
-        assert_not_equal(parsed_result, False)
-        assert_equal(80,len(parsed_result))
-        for result in parsed_result:
-            assert_equal(sorted((u'protocol', u'source', u'destination', u'srcport', u'destport', u'eni', u'action', u'packets', u'bytes', u'windowstart', u'windowend', u'version', u'account', u'flowlogstatus',u'envelope')), sorted(result.keys()))
-            assert_equal(sorted((u"logGroup",u"logStream",u"owner")),sorted(result['envelope'].keys()))
-
-class TestKVParser(object):
-    def setup(self):
-        """Setup before each method"""
-        # load config
-        self.config = load_config('test/unit/conf')
-        # load JSON parser class
-        self.parser_class = get_parser('kv')
-
-    def teardown(self):
-        """Teardown after each method"""
-        pass
-
-    def parser_helper(self, **kwargs):
-        data = kwargs['data']
-        schema = kwargs['schema']
-        options = kwargs['options']
-
-        kv_parser = self.parser_class(data, schema, options)
-        parsed_result = kv_parser.parse()
-        return parsed_result
-
-    def test_kv_parsing(self):
-        """Parse KV - 'key:value,key:value'"""
-        # setup
-        schema = {
-            'name': 'string',
-            'result': 'string'
-        }
-        options = {
-            'separator': ':',
-            'delimiter': ',',
-            'service': 'kinesis'
-        }
-        data = 'name:joe bob,result:success'
-
-        # get parsed data
-        parsed_data = self.parser_helper(data=data, schema=schema, options=options)
-
-        assert_equal(len(parsed_data), 1)
-        assert_equal(parsed_data[0]['name'], 'joe bob')
